add transcription hook
